requirements.txt
gradio==3.27.0
numpy==1.21.5
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.25.1
